Reconcile definitions with the headers that declare them
Seventy-two functions were defined with a signature their own header
contradicted, which no compiler ever saw because the files do not
include those headers. Fifty-three are now settled: filled-in return
types, the fixed-width spelling on both sides, and parameter names that
match.

The rest turned out to be load-bearing and are documented instead. The
text and sprite family stays unprototyped because retail calls
DrawText8x8 with three arguments in one place and four in another, and
GameQueueShadedSprite with nine of its ten; InterpolateTrackAngle is
called with one argument and reads the second out of $a1. The asset
installers keep void * because every caller hands them the raw asset
block. DecDCTin keeps two spellings because the FMV code and libpress
each own theirs.

DrawEnvPacketSource moves next to the DrawEnv whose head it describes.

+++ src/main/PAL/main/audio/stop_sound_slot_voice.c
@@ -1,4 +1,4 @@
 #include "common.h"
 #include "psyq/snd.h"
 /* Stop the effect voice for sound slot `slot` (hardware voice = slot + 0xE). */
-void StopSoundSlotVoice(s16 slot) { SsUtKeyOffV((s16)(slot + 14)); }
+void StopSoundSlotVoice(s32 slot) { SsUtKeyOffV((s16)(slot + 14)); }
